requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
dateutils==0.6.12
dnspython==2.7.0
Flask==3.1.0
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
mthrottle==0.0.2
nse==1.0.6
nsetools==1.0.11
numpy==2.1.3
packaging==24.2
pandas==2.2.3
praw==7.8.1
prawcore==2.4.0
pymongo==4.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
six==1.17.0
soupsieve==2.6
stocksnap==0.1.0
tzdata==2024.2
update-checker==0.18.0
urllib3==2.2.3
websocket-client==1.8.0
Werkzeug==3.1.3
